feat: Header와 연결한 Sidemenu 수정 후 기존 파일 삭제
