--- src/layout/SideMenuStyle.ts
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-export const SideContainer = styled.div`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 330px;
-  height: 100%;
-  background-color: #e1e1e1;
-  display: flex;
-  justify-content: space-between;
-`;
-
-export const SideMenuContents = styled.div``;
