+++ src/layout/style/SideMenuStyle.ts
@@ -14,17 +14,17 @@ export const MenuOverlay = styled.div`
     // backdrop-filter: blur(0.5px); 
 `;
 
-export const HomeSideContainer = styled.div`
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 330px;
-    height: 100%;
-    background-color: #e1e1e1;
-    display: flex;
-    justify-content: space-between;
+export const SideContainer = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 330px;
+  height: 100%;
+  background-color: #e1e1e1;
+  display: flex;
+  justify-content: space-between;
 `;
 
-export const HomeSideMenuContents = styled.div`
-    padding: 50px;
+export const SideMenuContents = styled.div`
+  padding: 50px;
 `;
